Rename configCtrl to configEditor
